+++ framework/model/src/test/java/is/codion/framework/model/ForeignKeyConditionModelTest.java
@@ -54,15 +54,15 @@ public class ForeignKeyConditionModelTest {
 		EntitySearchModel inSearchModel = conditionModel.inSearchModel();
 		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("SALES"));
 		inSearchModel.entity().set(sales);
-		Collection<Entity> searchEntities = conditionModel.operand().in().get();
+		Collection<Entity> searchEntities = conditionModel.operands().in().get();
 		assertEquals(1, searchEntities.size());
 		assertTrue(searchEntities.contains(sales));
 		Entity accounting = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("ACCOUNTING"));
 		List<Entity> salesAccounting = asList(sales, accounting);
 		inSearchModel.entities().set(salesAccounting);
-		assertTrue(conditionModel.operand().in().get().contains(sales));
-		assertTrue(conditionModel.operand().in().get().contains(accounting));
-		searchEntities = conditionModel.operand().in().get();
+		assertTrue(conditionModel.operands().in().get().contains(sales));
+		assertTrue(conditionModel.operands().in().get().contains(accounting));
+		searchEntities = conditionModel.operands().in().get();
 		assertEquals(2, searchEntities.size());
 		assertTrue(searchEntities.contains(sales));
 		assertTrue(searchEntities.contains(accounting));
@@ -77,23 +77,23 @@ public class ForeignKeyConditionModelTest {
 		EntitySearchModel equalSearchModel = conditionModel.equalSearchModel();
 		Entity sales = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("SALES"));
 		equalSearchModel.entity().set(sales);
-		Entity searchEntity = conditionModel.operand().equal().get();
+		Entity searchEntity = conditionModel.operands().equal().get();
 		assertSame(sales, searchEntity);
 		Entity accounting = CONNECTION_PROVIDER.connection().selectSingle(Department.NAME.equalTo("ACCOUNTING"));
 		equalSearchModel.entity().set(accounting);
-		assertSame(accounting, conditionModel.operand().equal().get());
+		assertSame(accounting, conditionModel.operands().equal().get());
 
 		equalSearchModel.entity().clear();
 
-		searchEntity = conditionModel.operand().equal().get();
+		searchEntity = conditionModel.operands().equal().get();
 		assertNull(searchEntity);
 
-		conditionModel.operand().equal().set(sales);
+		conditionModel.operands().equal().set(sales);
 		assertEquals("SALES", equalSearchModel.searchString().get());
 		sales.put(Department.NAME, "sales");
 		equalSearchModel.entity().set(sales);
 		sales.put(Department.NAME, "SAles");
-		conditionModel.operand().equal().set(sales);
+		conditionModel.operands().equal().set(sales);
 		assertEquals("SAles", equalSearchModel.searchString().get());
 	}
 }